// providers/types.ts
export interface Catalog {
  title: string;
  filter: string;
}